pyproject.toml
[tool.poetry]
name = "disp8210"
version = "0.1.0"
description = ""
authors = ["Lanterman <splen1011@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
django-debug-toolbar = "^6.0.0"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
python-dotenv = "^1.1.1"
coverage = "^7.10.4"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
